WinAPI.VK instead of 'ushort' in Dispatch()

+++ Mubox/Extensions/ExtensionManager.cs
@@ -158,7 +158,7 @@ namespace Mubox.Extensions
                 var client = new Extensibility.MuboxClientBridge(
                     (key) =>
                     {
-                        e.Client.Dispatch((ushort)key);
+                        e.Client.Dispatch(key);
                     },
                     (me) =>
                     {
